Split each frame in a different function, deinlining prototype
Splitting each frame in its own function required the code to be generated into
an internal list and then written out once generated, as the globals need to be
placed at the start of the file and are not known until the whole code has been
generated.

A better solution is to generate two files, one for globals and another for code,
so no temporary storage is needed and we start earlier to write to result files,
but this will have to wait until command-line parameters are accepted.

For deinlining, we build a histogram of each possible substring of each frame
and then substitute the one that gives highest compression, and iterate.

There are speed and memory drawbacks of this approach.
- For the speed, the histogram doesn't need to be fully rebuilt, only those
  frames that had any replacement done (and only the substrings intersecting
  the replaced, but this can get tricky to subtract incrementally if there were
  several replacements)
- For the memory, a trie should be used instead of replicating all the possible
  strings in a dict
- For both, an LZ approach could be used where the compression is done single-pass

glparse.py
- Changed to use global_decls and code_frames instead of direct stdout writes

trace.c
- Added stub for glMapBufferRange, needed by some traces.
- Removed draw(), this is now created by the Python script.

deinline.py
- Created. This takes the embedded trace or a trace.inc file and roughly creates
  the deinlined or "refactored" version (currently missing parameter and assignment
  handling)

Tests:
Ran kipo, bmk_bitmap traces on Tegra3 and Adreno330
Deinlined kipo, gtavc (this triggers out of memory).