Added parameter handling to deinline.py
When a block of code is factored out, the set of actual parameters are calculated
as all the parameters to all the functions inside that block.
Once all the callers for that block have been factored out, the formal parameters
are calculated as those not in the intersection of the actual parameters of each
caller.
The intersection is placed as immediate values in the factored code block and
eliminated from the actual parameters of each call site, as it's constant across
all callers (unless those parameters are local variables, in which case they are
still passed as parameters)

main.c
- Enabled stop_motion by default

trace.c
- Added getAssetBuffer and openAssetAndGetBuffer (only the latter is used in order
  to avoid aliasing issues when deinlining a piece of code that writes to a pointer
  variable and uses it later).

build.bat
- Added deinlining step

build.sh
- Added deinlining step
- Fixed -buildfile not working on Linux, needs to switch to the directory where
  the build file is.

deinline.py
- Added actual and formal parameter calculation and handling.
- Fixed off by one errors caused by xrange() not including the final value
- Removed sample frames, as now function prototypes and globals are also needed
- Improved the parser so it does the right thing with globals, function prototypes
  and non-function code inside functions.
- Added deriving parameter type from the mangled parameter name. This is necessary
  in order to generate the function prototypes.
  There are interesting aliasing issues that are worked around by never generating
  code that hit them (see comments in the trace.c regarding deilining a block of
  code that initializes a pointer and then uses that pointer variable).

glparse.py
- Added name mangling to all variables so their type can be discerned by the
  deinliner.
- Removed extra empty frame generation and missing last frame call.
- Added ignoring contexts and disabled eglMakeCurrent handling, as it was broken
  anyway for multithreaded.
- Fixed RenderBuffer to Renderbuffer and such.
- Refactored asset management, now used for shaders, floats, etc.

Tests:
Ran kipo, sample_frames, gtavc traces though deinliner. gtavc with all frames
still triggers out of memory.
Ran 200 frames of gtavc trace without deinlining, mostly works, but it lacks the
texture compression.
Ran kipo deinlined on Tegra3 (200 frames)