Deinline aliased parameter coalescing
deinlined traces were crashing with NULL pointer accesses on the Android device.
This was due to a new incarnation of the aliasing problem that was workarounded
at the time by joining openAsset and getBuffer into openAssetAndGetBuffer.

Turns out there are more cases where that aliasing bug happens, eg when getting
a pointer to an asset and then using it in glTexImage2D or glVertexAttribPointerData

When deinlining that code, the address of the pointer and the pointer itself
would be passed as different parameters, which broke the aliasing, caused
 the address of the pointer to be updated but not the function parameter, so NULL
pointers were passed while inside the function.

The current solution to that problem is to coalesce the address of the pointer
and the pointer into a single parameter, so inside the body of the refactored
function only the address and its dereference are used.

This has the limitation that it doesn't support one caller passing aliased
parameters and another calling passing non-aliased parameters.
This could be dealt with in different ways (removing the aliased or the non-aliased
from the refactoring, etc) but currently an exception is thrown.

activity\jni\trace.c
- Added exit in case of error.

build.bat
- Enabled deinlining

build.py
- Added logging to deinline when scriptine is in verbose mode.

deinline.py
- Added code to coalesce aliased parameters.

Tests:

complex.c
- Changes due to coalescing aliased parameters (removed formal & actual parameters).

params_bug_aliasing.c
- Added test to check for different types of aliasing.

build.bat for deinlined SonicDash, compression ratio is 35x, works fine on Tegra3
(used to crash)

tests/make.bat

Ran 19 tests in 26.646s

OK